make netifaces dependency optional because it is not easy to install on all systems
