+++ nicegui/welcome.py
@@ -1,0 +1,32 @@
+from typing import List
+from .globals import optional_features
+import os
+
+try:
+    import netifaces
+    optional_features.append('netifaces')
+except ImportError:
+    pass
+
+
+def get_all_ips() -> List[str]:
+    if not 'netifaces' in optional_features:
+        return []
+    ips = []
+    for interface in netifaces.interfaces():
+        try:
+            ips.append(netifaces.ifaddresses(interface)[netifaces.AF_INET][0]['addr'])
+        except KeyError:
+            pass
+    return ips
+
+
+def print_message():
+    host = os.environ['NICEGUI_HOST']
+    port = os.environ['NICEGUI_PORT']
+    ips = set(get_all_ips() if host == '0.0.0.0' else [])
+    ips.discard('127.0.0.1')
+    addresses = [(f'http://{ip}:{port}' if port != '80' else f'http://{ip}') for ip in ['localhost'] + sorted(ips)]
+    if len(addresses) >= 2:
+        addresses[-1] = 'and ' + addresses[-1]
+    print(f'NiceGUI ready to go on {", ".join(addresses)}', flush=True)
